Consolidate the atespace controlapi handlers into one file
No declaration added, removed, or renamed, import sets
unchanged, and every non-header line identical to what it replaced.
